% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmatrix.R
\name{Pmatrix}
\alias{Pmatrix}
\title{Matriz de correlaciones parciales.}
\usage{
Pmatrix(x)
}
\arguments{
\item{x}{matriz de datos que se va a analizar.}
}
\value{
Matriz con las correlaciones parciales entre las diferentes variables de la matriz de datos
}
\description{
Función que nos va a dar como resultado las correlaciones parciales
entre las diferentes variables de la matriz de datos
}
